Remove tests/common_unit_test.py
