+++ tests/test_local_copy_paste.py
@@ -5,7 +5,7 @@ import sys
 
 import pytest
 
-from .common_unit_test import FILE_CONTENT, UnitTestCase
+from .common import FILE_CONTENT, UnitTestCase
 
 
 class TestLocalCopyPaste(UnitTestCase):
