+++ tests/test_local_share_move_folders.py
@@ -5,7 +5,7 @@ import shutil
 from mock import patch
 
 from nxdrive.engine.watcher.remote_watcher import RemoteWatcher
-from .common_unit_test import UnitTestCase
+from .common import UnitTestCase
 
 wait_for_security_update = False
 src = None
